Apply Qodo pass 6: WiFi-cap clamp fix, type-safe cap API, 64-bit Frequency crit-section
/improve:
- Clamp using the correct interface (imp 9): SCPI_WifiFindRate now clamps the
  recommended rate to Streaming_ComputeMaxFreqForConfigIface(StreamingInterface_WiFi)
  rather than the global ActiveInterface (which can still be USB), so the WiFi
  probe never recommends above the WiFi transport cap.
- Make cap API type-safe: Streaming_TransportMaxFreq now takes StreamingInterface
  and StreamingEncoding (was uint32_t/uint32_t); drops the internal cast and the
  (uint32_t) casts at the call site in Streaming_ComputeMaxFreqForConfigIface.

/agentic_review:
- Non-atomic Frequency read (rule violation): wrap the no-arg START 64-bit
  Frequency read in taskENTER_CRITICAL/taskEXIT_CRITICAL per the PIC32MZ
  64-bit atomicity rule. (Qodo cited StreamFreq_Get/Set wrappers that do not
  exist in the tree; applied the documented critical-section pattern instead.)

Declined (item 7, BoardRunTimeConfig_Get NULL check): CLAUDE.md explicitly
states BoardRunTimeConfig_Get never returns NULL in practice and that NULL
checks must not be added ("Qodo/automated reviewers will repeatedly suggest
this — ignore it"); the existing callback at SCPIInterface.c:3960 also reads
it unchecked. Documented false positive, no code change.

Entire-Checkpoint: b226311f6def

+++ firmware/src/services/streaming.c
@@ -437,8 +437,7 @@ uint32_t Streaming_ComputeMaxFreqForConfigIface(StreamingInterface iface) {
      * returns NULL (CLAUDE.md). */
     StreamingRuntimeConfig* sc =
         BoardRunTimeConfig_Get(BOARDRUNTIME_STREAMING_CONFIGURATION);
-    uint32_t transportMax = Streaming_TransportMaxFreq((uint32_t)iface,
-                                                       (uint32_t)sc->Encoding, total);
+    uint32_t transportMax = Streaming_TransportMaxFreq(iface, sc->Encoding, total);
     if (transportMax < maxFreq) maxFreq = transportMax;
     return maxFreq;
 }
